:sparkles:  Add the `unique` option for Spherical
This option guarantee the Spherical Coordinate will _always_ be unique (r ∈ [0, ∞), θ ∈ [0, π] and φ ∈ [0, 2π)). This is not active by default because, when it is active,
`s=Spherical([1,np.pi,0], unique=True);s.theta+=np.pi/2;s.theta-=np.pi/2;` makes `s != Spherical([1,np.pi,0])`. This is NOT a bug, but a consequence.